requirements.txt
python-dotenv
langchain-core
langchain-community
ddgs
pydantic
langchainhub
langchain-groq